Move public-facing operators to nodes.hpp and remove operators header.
