+++ cpp/include/cudf/ast/detail/linearizer.hpp
@@ -17,7 +17,6 @@
 
 #include <cudf/ast/detail/operators.hpp>
 #include <cudf/ast/nodes.hpp>
-#include <cudf/ast/operators.hpp>
 #include <cudf/scalar/scalar_device_view.cuh>
 #include <cudf/table/table_view.hpp>
 #include <cudf/types.hpp>
