+++ cpp/src/ast/linearizer.cpp
@@ -15,7 +15,6 @@
  */
 #include <cudf/ast/detail/linearizer.hpp>
 #include <cudf/ast/nodes.hpp>
-#include <cudf/ast/operators.hpp>
 #include <cudf/scalar/scalar.hpp>
 #include <cudf/scalar/scalar_device_view.cuh>
 #include <cudf/table/table_view.hpp>
