+++ cpp/tests/join/conditional_join_tests.cu
@@ -15,7 +15,6 @@
  */
 
 #include <cudf/ast/nodes.hpp>
-#include <cudf/ast/operators.hpp>
 #include <cudf/column/column_view.hpp>
 #include <cudf/join.hpp>
 #include <cudf/table/table_view.hpp>
